refector : 문자열 계산기 피드백 이후 수정
